requirements.txt
requests~=2.27.1
awscli
flask~=1.1.2
selenium~=3.141.0
Pillow~=8.4.0
names~=0.3.0